Finished FilterDropdown Component
