+++ src/components/common/FilterDropdown.jsx
@@ -1,42 +1,28 @@
-import React, { useState } from 'react';
 import Dropdown from 'react-bootstrap/Dropdown';
-import Form from 'react-bootstrap/Form';
+import CustomMenu from './FilterDropdownMenu';
 
-// eslint-disable-next-line react/display-name
-const CustomMenu = React.forwardRef(
-  ({ children, style, className, 'aria-labelledby': labeledBy }, ref) => {
-    const [value, setValue] = useState('');
-
-    return (
-      <div ref={ref} style={style} className={className} aria-labelledby={labeledBy}>
-        <Form.Control
-          autoFocus
-          className="mx-3 my-2 w-auto"
-          placeholder="Type to filter..."
-          onChange={(e) => setValue(e.target.value)}
-          value={value}
-        />
-        <ul className="list-unstyled">
-          {React.Children.toArray(children).filter(
-            (child) => !value || child.props.children.toLowerCase().startsWith(value.toLowerCase())
-          )}
-        </ul>
-      </div>
-    );
-  }
-);
-function FilterDropdown() {
+function FilterDropdown({ currElement, data, callbackFunc }) {
   return (
-    <Dropdown>
-      <Dropdown.Toggle id="dropdown-custom-components">Custom toggle</Dropdown.Toggle>
+    <Dropdown onSelect={(eventKey) => callbackFunc(eventKey)}>
+      <Dropdown.Toggle id="dropdown-custom-components">{currElement}</Dropdown.Toggle>
 
       <Dropdown.Menu as={CustomMenu}>
-        <Dropdown.Item eventKey="1">Red</Dropdown.Item>
-        <Dropdown.Item eventKey="2">Blue</Dropdown.Item>
-        <Dropdown.Item eventKey="3" active>
-          Orange
-        </Dropdown.Item>
-        <Dropdown.Item eventKey="1">Red-Orange</Dropdown.Item>
+        {data.map((value, index) => {
+          if (value === currElement) {
+            return (
+              // eslint-disable-next-line react/no-array-index-key
+              <Dropdown.Item key={index} eventKey={value} active>
+                {value}
+              </Dropdown.Item>
+            );
+          }
+          return (
+            // eslint-disable-next-line react/no-array-index-key
+            <Dropdown.Item key={index} eventKey={value}>
+              {value}
+            </Dropdown.Item>
+          );
+        })}
       </Dropdown.Menu>
     </Dropdown>
   );
